+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ## 🚀 Feature
 - Added multiline support for `AndesSwitch` | Author: [@juanunzu](https://github.com/juanunzu)
 - Fresco version updated to 2.2.0 | Author: [@eportasML](https://github.com/eportasML)
+- AndesSwitch package rename to `switchandes` to be java compatible | Author: [@eportasML](https://github.com/eportasML)
 
 # v3.20.0
 ## 🚀 Feature